:recycle: Move formatAmountBasedOnLocale to its on file
