+++ src/components/delegatesListView/delegateRow.js
@@ -8,7 +8,7 @@ import AccountVisual from '../accountVisual';
 import LiskAmount from '../liskAmount';
 import routes from '../../constants/routes';
 import i18n from '../../i18n';
-import { formatBasedOnLocale } from '../../utils/lsk';
+import { formatAmountBasedOnLocale } from '../../utils/formattedNumber';
 
 const setRowClass = (voteStatus) => {
   if (!voteStatus) {
@@ -109,7 +109,7 @@ class DelegateRow extends React.Component {
           {` ${this.props.t('LSK')}`}
         </div>
         <div className={`${columnClassNames.productivity} delegate-productivity`}>
-          {`${formatBasedOnLocale({ value: productivity, locale: i18n.language })} %`}
+          {`${formatAmountBasedOnLocale({ value: productivity, locale: i18n.language })} %`}
         </div>
         <div className={`${columnClassNames.voteWeight} ${styles.weight} vote-weight`}>
           <LiskAmount val={data.vote} />
